Refactoring the Encrypt functions to a new class
