+++ development/actuators/demo-system-service/src/main/java/eubr/atmosphere/tma/actuator/filters/DecryptFilter.java
@@ -1,26 +1,13 @@
 package eubr.atmosphere.tma.actuator.filters;
 
-import java.io.DataInputStream;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.nio.charset.Charset;
 import java.security.InvalidKeyException;
-import java.security.KeyFactory;
 import java.security.NoSuchAlgorithmException;
 import java.security.PrivateKey;
 import java.security.PublicKey;
-import java.security.Signature;
 import java.security.SignatureException;
-import java.security.interfaces.RSAPrivateKey;
-import java.security.interfaces.RSAPublicKey;
-import java.security.spec.InvalidKeySpecException;
-import java.security.spec.PKCS8EncodedKeySpec;
-import java.security.spec.X509EncodedKeySpec;
 import java.util.Base64;
 
-import javax.crypto.Cipher;
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
 import javax.servlet.FilterConfig;
@@ -36,6 +23,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
+import eubr.atmosphere.tma.actuator.crypto.KeyManager;
 import eubr.atmosphere.tma.actuator.wrappers.HttpServletRequestWritableWrapper;
 import eubr.atmosphere.tma.actuator.wrappers.HttpServletResponseCopier;
 
@@ -44,12 +32,6 @@ import eubr.atmosphere.tma.actuator.wrappers.HttpServletResponseCopier;
 public class DecryptFilter implements Filter {
     
     private static final Logger LOGGER = LoggerFactory.getLogger(DecryptFilter.class);
-
-    public static final String ALGORITHM = "RSA";
-    public static final Charset UTF_8 = Charset.forName("UTF-8");
-    public static final String SIGNATURE_ALGORITHM = "SHA1withRSA";
-    public static final String SECRET = "!@#$MySecr3tPasSw0rd";
-    public static final String SYMMETRIC_KEY_ALGORITHM = "AES";
 
     // This is the current test:
     // curl --header "Content-Type: text/plain" --request POST --data-binary "@encrypted-message" http://localhost:8080/securePOC/act
@@ -65,8 +47,8 @@ public class DecryptFilter implements Filter {
         HttpServletRequest req = (HttpServletRequest) servletRequest;
         LOGGER.info("Logging Request {} : {}", req.getMethod(), req.getRequestURI());
 
-        PrivateKey privateKey = getPrivateKey("/home/virt-atm/Documents/priv-key-again");
-        byte[] bodyBytes = decrypt(getBody(req), privateKey);
+        PrivateKey privateKey = KeyManager.getPrivateKey("/home/virt-atm/Documents/priv-key-again");
+        byte[] bodyBytes = KeyManager.decrypt(getBody(req), privateKey);
         String decryptedData = new String(bodyBytes);
         LOGGER.info(decryptedData);
 
@@ -98,95 +80,14 @@ public class DecryptFilter implements Filter {
     private void encryptResponse(ServletResponse servletResponse, PrivateKey privateKey,
             String plainResponse) throws InvalidKeyException, IOException, NoSuchAlgorithmException, SignatureException {
 
-        byte[] signedResponseByteArray = sign(plainResponse, privateKey);
+        byte[] signedResponseByteArray = KeyManager.sign(plainResponse, privateKey);
         String signedResponse = Base64.getEncoder().encodeToString(signedResponseByteArray);
         LOGGER.info("signedResponse: " + signedResponse);
 
-        PublicKey publicKeyExecutor = getPublicKey("/home/virt-atm/Documents/pub-key-execute");
-        servletResponse.getWriter().write(Base64.getEncoder().encodeToString(encrypt(plainResponse.getBytes(), publicKeyExecutor)));
+        PublicKey publicKeyExecutor = KeyManager.getPublicKey("/home/virt-atm/Documents/pub-key-execute");
+        servletResponse.getWriter().write(Base64.getEncoder().encodeToString(KeyManager.encrypt(plainResponse.getBytes(), publicKeyExecutor)));
         servletResponse.getWriter().write("\n");
         servletResponse.getWriter().write(signedResponse);
-    }
-
-    //The method that signs the data using the private key that is stored in keyFile path
-    public byte[] sign(String data, PrivateKey keyFile)
-            throws NoSuchAlgorithmException, InvalidKeyException, SignatureException {
-        Signature privateSignature = Signature.getInstance(SIGNATURE_ALGORITHM);
-        privateSignature.initSign(keyFile);
-        privateSignature.update(data.getBytes(UTF_8));
-
-        byte[] signature = privateSignature.sign();
-        return signature;
-    }
-
-    public static boolean verify(String plainText, String signature, PublicKey publicKey)
-            throws NoSuchAlgorithmException, InvalidKeyException, SignatureException {
-        Signature publicSignature = Signature.getInstance(SIGNATURE_ALGORITHM);
-        publicSignature.initVerify(publicKey);
-        publicSignature.update(plainText.getBytes(UTF_8));
-
-        byte[] signatureBytes = Base64.getDecoder().decode(signature);
-
-        return publicSignature.verify(signatureBytes);
-    }
-
-    private PrivateKey getPrivateKey(String filenamePrivKey) {
-        try {
-            File privKeyFile = new File(filenamePrivKey);
-
-            // read private key DER file
-            DataInputStream dis = new DataInputStream(new FileInputStream(privKeyFile));
-            byte[] privKeyBytes = new byte[(int)privKeyFile.length()];
-            dis.read(privKeyBytes);
-            dis.close();
-
-            KeyFactory keyFactory = KeyFactory.getInstance(ALGORITHM);
-
-            // decode private key
-            PKCS8EncodedKeySpec privSpec = new PKCS8EncodedKeySpec(privKeyBytes);
-            RSAPrivateKey privKey = (RSAPrivateKey) keyFactory.generatePrivate(privSpec);
-
-            return privKey;
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (InvalidKeySpecException e) {
-            e.printStackTrace();
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-        }
-
-        return null;
-    }
-
-    private PublicKey getPublicKey(String filenamePubKey) {
-        try {
-            File pubKeyFile = new File(filenamePubKey);
-
-            DataInputStream dis = new DataInputStream(new FileInputStream(pubKeyFile));
-            byte[] pubKeyBytes = new byte[(int)pubKeyFile.length()];
-            dis.readFully(pubKeyBytes);
-            dis.close();
-
-            KeyFactory keyFactory = KeyFactory.getInstance(ALGORITHM);
-
-            // decode public key
-            X509EncodedKeySpec pubSpec = new X509EncodedKeySpec(pubKeyBytes);
-            RSAPublicKey pubKey = (RSAPublicKey) keyFactory.generatePublic(pubSpec);
-
-            return pubKey;
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (InvalidKeySpecException e) {
-            e.printStackTrace();
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-        }
-
-        return null;
     }
 
     @Override
@@ -203,34 +104,4 @@ public class DecryptFilter implements Filter {
 
         return null;
     }
-
-    private static byte[] decrypt(byte[] text, PrivateKey key) {
-        byte[] dectyptedText = null;
-        try {
-            // get an RSA cipher object and print the provider
-            final Cipher cipher = Cipher.getInstance(ALGORITHM);
-
-            // decrypt the text using the private key
-            cipher.init(Cipher.DECRYPT_MODE, key);
-            dectyptedText = cipher.doFinal(text);
-
-         } catch (Exception ex) {
-             ex.printStackTrace();
-         }
-         return dectyptedText;
-    }
-
-    private static byte[] encrypt(byte[] text, PublicKey key) {
-        byte[] cipherText = null;
-        try {
-            // get an RSA cipher object and print the provider
-            final Cipher cipher = Cipher.getInstance(ALGORITHM);
-            // encrypt the plain text using the public key
-            cipher.init(Cipher.ENCRYPT_MODE, key);
-            cipherText = cipher.doFinal(text);
-          } catch (Exception e) {
-            e.printStackTrace();
-          }
-          return cipherText;
-   }
 }
